+++ buildpacks/jvm-function-invoker/CHANGELOG.md
@@ -3,6 +3,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+* Upgrade `libcnb` to `0.5.0` and `libherokubuildpack` to `0.5.0`.
 
 ## [0.6.0] 2022/01/05
 * Switch to BSD 3-Clause License
@@ -12,7 +13,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.5.5] 2021/10/19
 
 ## [0.5.4] 2021/09/30
-
 * Updated function runtime to `1.0.3`
 
 ## [0.5.3] 2021/09/29